# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/admin/add": {
            "post": {
                "description": "Add an API key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication token (e.g., Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key ID for authorization.",
                        "name": "Authorization-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key on which to perform the action.",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key ID on which to perform the action.",
                        "name": "X-API-Key-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "scope",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "permissions",
                        "name": "permissions",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/delete": {
            "post": {
                "description": "Delete an API key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication token (e.g., Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key ID for authorization.",
                        "name": "Authorization-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key on which to perform the action.",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key ID on which to perform the action.",
                        "name": "X-API-Key-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/edit": {
            "post": {
                "description": "Edit an API key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication token (e.g., Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key ID for authorization.",
                        "name": "Authorization-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key on which to perform the action.",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API key ID on which to perform the action.",
                        "name": "X-API-Key-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "scope",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "permissions",
                        "name": "permissions",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}